Rewrite initialization to use copy construction instead of assignment.
This is a work-in-progress.

The initial change causes an unresolved call, attempting to coerce an int(32) to int(64).

I suspect that assignment has magical powers w.r.t. to coercion that would have to be
transferred to constructors for this to work.  This represents a major design change, and
a change in the compiler semantics, so warrants a bit of discussion before I proceed.