Fix minor problem in _defaultOf() implementation and update no-use-record.bad
After the initial merge of defaultOfDominates, we got a .bad file mismatch on the
no-use-record future.  The error message says we are trying to convert nil to (undefined)
type R, but nil is the default value for a class not a record, so clearly something is
wrong.

There was the _defaultOf() function
```inline proc _defaultOf(type t) {
```  select t {
```    when memory_order {
```      return _defaultOfMemoryOrder();
```    }
```    otherwise {
```      return nil:t;
```    }
```  }
```}
The works OK for memory_order thingies, but casts a much wider net.  I narrowed it so it
addresses only the memory_order case, and then had to supply a missing definition for
_ddata, because the "no object" pragma makes a class type fail the isClassType() test (!).

Then I updated the .bad file appropriately.  That doesn't solve the error completely (or
at all, actually), but it does make the _defaultOf() implementation more accurate.