Bridged out the code in fixupDestructors() that handles syncs and removable autoDestroys specially.
The test case types/records/sungeun/classInRecord.chpl was failing because we are calling
the destructor for a class directly (bypassing chpl__autoDestroy()).

The idea of a removable autoDestroy does not really fit into the new model.  We will add
autoDestroy call where they are needed.  If all goes well, autoDestroy will be called
uniformly for all destruction, and every one will trivially call the corresponding
destructor.  When that is true, then the crufty concept of autoDestroy can be removed, and
we will just call the destructor directly.