Modified BasicBlock::restart() to eliminate empty blocks with no predecessors.
If restart() is called and the current block is empty and has no predecessors, it is
reused.  This solves the problem where a conditional "then" or "else" clause containing a
goto statement (like "break" or "continue") creates an empty block with no predecessors.

The goto statement has to start a new block, because normally blocks are chained
together in sequence.  A goto statement ends the current block and sets its successor to the
block that contains the target label.  The new block initially has no predecessors, but a
predecessor can be added if the block begins with a label.

If there is a block with no predecessors that is not empty, this means there is code that
is unreachable.  That case should probably be handled earlier in parsing, so I added code
to verifyBasicBlocks to check for this.  We currently violate this in how we do param
folding of conditional expressions, so I left it ifdef'd out.

Made block verification depend on the fVerify flag.

Added some comments to bb.cpp.