Remove unreachable blocks in pruneResolvedTree().
Ownership flow analysis was getting the wrong answer due to unreachable blocks remaining
in the tree after resolution.  The problem is caused when a param "if" with an else clause
is folded: The condition is removed so the "then" clause is always reached, but the else
clause is not removed from the tree.

Ownership flow analysis starts with nothing owned at the top of the else block.  This
usually results in an empty OUT set, and this clears every bit in the IN set of the
following block.  Backward propagation also clears the OUT set of the (last block of the)
"then" clause.  This finally causes every owned variable to be destroyed at the end of the
"then" clause, which in some cases is premature.

###DETAILS
+++ b/compiler/include/astutil.h
+++ b/compiler/AST/astutil.cpp
Added collect_stmts_postorder_STL(BaseAST* ast, std::vector<Expr*>& stmts)

+++ b/compiler/include/bb.h
+++ b/compiler/AST/bb.cpp
Factored the block removal out of removeEmptyBlocks() and made it public, so it could be
used in unreachable block removal in functionResolution as well.

@@ -51,6 +51,7 @@ void collect_top_asts_STL(BaseAST* ast, std::vector<BaseAST*>& asts);
 void collectExprs(BaseAST* ast, std::vector<Expr*>& exprs);
 void collect_stmts(BaseAST* ast, Vec<Expr*>& stmts);
 void collect_stmts_STL(BaseAST* ast, std::vector<Expr*>& stmts);
+void collect_stmts_postorder_STL(BaseAST* ast, std::vector<Expr*>& stmts);
 void collectDefExprs(BaseAST* ast, Vec<DefExpr*>& defExprs);
 void collectDefExprsSTL(BaseAST* ast, std::vector<DefExpr*>& defExprs);
 void collectCallExprs(BaseAST* ast, Vec<CallExpr*>& callExprs);

+++ b/compiler/optimizations/deadCodeElimination.cpp
The fixup in deadCodeElimination of conditionals from which blocks have been removed is
code that has never executed.  I rewrote it so it tests only one dimension per conditional
and revised the block manipulation so it actually works.

+++ b/compiler/resolution/functionResolution.cpp
Added removeDeadBlocks to pruneResolvedTree().
TODO: See if we can use the deadBlockElimination() routine that already exists.