Treat assignment to the return value variable as consumption.
Do not treat the return primitive as consumption any more.

This change was necessary because there are multiple assignments to the return value
variable.  That was causing otherwise unrelated alias cliques to be joined together.  And
then in some cases, the leader of a clique ws being autoDestroyed even when it was not
owned.

For example, in
  proc +(s: string_rec, ref cs: c_string_copy) {
there are multiple returns, such as
    if cs == _defaultOf(c_string_copy) then return s;
Internally, this bound s to the return value variable (RVV), but at that point in the code
s refers to an argument and is therefore (by default) unowned.  In later code, there is a
"ret" variable that is created, and since it is default-constructed it is by definition
owned.  When it is assigned to the RVV, it sets the ownership state to true, and that
value propagates to the end of the function.  AMM sees that the clique is owned and
inserts an autodestroy on s.

Two problems: the ownership state of the clique is ambiguous at the end of the routine,
and the wrong variable gets autoDestroyed.  Treating assignment to the RVV as consumption
solves this problem.

There is another approach which is more precise but more costly.  This involves
maintaining the aliasing state exactly throughout the program.  Ensuring that the
representation is in single-assignment form probably eliminates the need for this
analysis.

+++ b/compiler/passes/insertAutoCopyAutoDestroy.cpp
In isDestructorArg() now handle the case where an autodestroy calls another autodestroy.
We also shouldn't call autocopy on the field args there.
Added initInSets, to not autoCopy the meme argument.  This prevents infinite recursion in
the non-method version of the compiler-supplied all-fields constructor.  When constructors
are all methods, this special code can be removed.