Add initCopy routine for types supporting privatization.
This is part of the rework to support privatized objects without creating
network traffic due to ref count updates.  Overall, this is an optimization
that assumes that privatized objects are shared and leaked.

-------------------

This patch in particular:

When a privatized object is created, we need to set its ref count to 1, under
the assumption that it is still treated as a ref counted object in some places
in the code.  The initial ref count of one will prevent it from ever being
deleted, even if a client causes the ref count to go away from and return to
the current ref count value.  (Returning to a ref count of zero causes a
ref-counted object to be deleted.)

TODO: This is an intermediate solution.  After reference counting is reworked
so that it is disabled in privatized versions of these objects, then the need
to set an initial ref count can be removed, along with all of the initCopy
routines added here.