First cut at getting wide strings working.
In resolveIntents(), if the base type of an argument is dtString and we resolve
it to "const ref" intent, then change the actual type of the argument from
string to _ref(string).

Rewrote replaceVarUsesWithFormals() to insert derefs on function calls as
needed.  If we treat the type of a string passed by blank intent to a function
as being "const ref" and make its base type _ref(string), then inside that
function, we may need a dereference to make the type of the actual match the
type of the formal one level down.

Widen string literals, because they refer to literals allocated back on locale 0.

In codegen, expand the special case -- to not codegen a string literal as a
literal string -- to include heap and wide types based on dtString.