Fix handling of c_ptr in codegen.
c_ptr is a value type, but it should be handled like a pointer in codegen.

I added special-purpose code to do this by keying off the type name.

TODO: This approach is rather ugly, but at least it works.  A better approach
would be to treat types like uint8_t as extern record types, and then ensure
that primitives like PRIM_CHPL_COMM_GET expect them to be passed by ref.
Codegen will see that the type is a ref type and generate the correct code
without the need for special-casing.