Postpone insertion of reference temps.
Reference temps were being inserted as early as possible, to establish the
invariant that formal and actual types match exactly.  Formerly, reference
temps were added as part of the callDestructors pass, so up to that point,
formals and actuals only necessarily match in their base types (ignoring ref
level).

However, this early insertion of reference temps was fooling the conditional at
line 866 of lowerIterators.cpp, so the required transformation was not applied.
This meant that in iterators where the yielded object is a record (returned by
value), its initial value was established by assignment.  But that breaks the
rules of AMM, since assignment can only be legally applied to an object that is
already initialized.  Some test cases involving iterators returning
strings-as-records were failing as a result:

    modules/standard/FileSystem/lydia/copyTree/copying.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/globberator-par.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/findfiles.0-2.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/globberator-zip.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/findfiles.0-4.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/findfiles-par.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-4.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-5.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-6.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-7.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-8.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-9.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/globberator.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/testboth.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk-par.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-1.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-2.exec.out.tmp:1:
    modules/standard/FileSystem/filerator/bradc/walk.0-3.exec.out.tmp:1:
    modules/standard/Path/lydia/realpath/fileNotValid.exec.out.tmp:1:

This change allows the transformation to work and those tests to pass.

This solution is not ideal.  Technically, the assignment operation is incorrect
in that context and should never be inserted in the first place.  There are a
number of advantages to rewriting return statements to they initialize the
return value with the return expression rather than default-initializing the
return value variable and then updating it with assignment.
In addition, the establishment and maintenance of invariants makes the compiler
implementation more flexible and easier to maintain going forward.